Initial codebase check-in
Description:
The initial core code is laid out in this
particular check-in. I still want to make a lot
of changes as far as structure / readability
goes but I wanted to put this out first.

Also a few tweaks have been made to the
documentation README.

Note: Several pieces of code have been commented
      out just to get the build passing; it will
      be finished in the immediate future.