Repairing TOML logic and general code expansion
Description:
The reference to the TOML was failing because of
a naming mismatch - it has been fixed. I've also
re-oreder the structs for the TOML parsing to be
a little more in line with what I've seen in the
BurntSushi/toml examples.

Several pieces of GitHub logic have been
adjusted (some as a result of the above
changes), both for clarity and code correctness.

Lastly, there are several large swaths of
unfinished Slack logic - this is partially due
to the API lacking some features (e.g. user name
to user ID helper methods) and partially due to
some waffling on my part to choose a specific
token / app structure for the service. This
still merits consideration.

Note: Code has been formatted correctly.
Note: Known compile errors (e.g. variable
      declared but not used).
Note: Several code pieces are commented out
      (WIP).