Restructure configuration setup
Description:
The way it had been set up previously was redundant (at least for the token
info) and now it should be a little clearer - there is a "shared" token for the
GitHub functionality. These "master" variables will serve as an override as
well (or at least as it pertains to channel variables).

There is probably going to need to be a helper function to sort out which
channels the message should be posted to (since things are currently stored as
slices and not maps) but we'll see.

Note: Building without errors.