Add logic for map references
Description:
These will make it more efficient in looking up what channels the message
should be posted into (there may be a more efficient architecture but this is
what I'm working with currently).

Additionally, there is some initial handler code being re-worked but currently
blocked out.

Lastly, I've included a test file; I'm not so sure a unit test is entirely
necessary, but I can't help but feel as though I'm betraying TDD in not
following it in this case (despite it being a bit late to being TDD). It may
ultimately be removed if it is found to be overkill.

Note: Build passing.