Outline project rework structure
Description:
This is an initial check-in to begin reformatting the project into
something (hopefully) cleaner and better.

I've made a separate directory to hold the bulk of the server code
and to allow for unit testing (which I'm obsessed with keeping up
with :grimacing:). The main.go file will just import and start the
server (most likely).

I'm not entirely sure that I need to have the request method check
inside the mainHandler but I've included it for now.

Note: WIP check-in
Note: One unit test (in a table) is implemented and passing
Note: This check-in include in-code comments and notes