Expand styling and content
Description:
I'm trying to keep the landing page as simple as possible; I'd like
to have something in place that can be a placeholder for status
updates or something. It may not ultimately be super value added but
I feel like it's a good exercise.

Note: WIP check-in
Note: Specific content will change as well as some styling