Gut and reorder all server/content files
Description:
I was going about things too complicated so now everything lives
inside main/_test.go files with the rest of the actual landing page
website content in the site/ directory. This should be as simple as
it can be made for a small app.

Note: WIP check-in
Note: No unit tests have been implemented yet