Implement simple incoming webhooks
Description:
I've geared back the immediate-term goals of the project. Instead of
focusing on a lot of features, I'd like to make it as lightweight as
possible and be able to have users deploy it to AWS Lambda fairly
quickly and easily.

To that end, I'll only be using incoming webhook features on Slack
(I still need to evaluate the security of including them in the
config.json file but I expect it to be no less risky than an
actual token - perhaps less so).

Note: WIP initial functionality check-in
Note: Posts correctly into the test channel